Parse HTTP response for rel="me" Link headers
Resolves #127.

To accomplish this, I swapped out the MicroMicro gem for Nokogiri and
nokogiri-html-ext with the addition of the link-header-parser gem.

The dependency changes cut back on overall total number of gems. Yay!

The search action is updated now to parse HTTP Link headers for rel="me"
URLs and combines that with rel="me" URLs parsed from the response body.

For URLs using only in-page rel="me" links, the results should be the
exact same.